LayoutTests/http/tests/referrer-policy-img/origin/cross-origin-http.https-expected.txt

CONSOLE MESSAGE: The page at https://127.0.0.1:8443/referrer-policy-img/origin/cross-origin-http.https.html was allowed to display insecure content from http://127.0.0.1:8000/referrer-policy/resources/image.py?expected=https://127.0.0.1:8443/.

Tests the behavior of origin referrer policy when cross origin from HTTPS to HTTP.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS referrer is "https://127.0.0.1:8443/"
PASS successfullyParsed is true

TEST COMPLETE
